Finished permissions with delete permissions based on user and per

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require("jsonwebtoken");
 const { randomBytes } = require("crypto");
 const { promisify } = require("util");
 const { transport, makeANiceEmail } = require("../mail");
+const { hasPermission } = require("../utils");
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -44,11 +45,20 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find item
-    const item = await ctx.db.query.item({ where }, `{id, title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
     // 2. Check if they own the item or permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+
+    if (!ownsItem) throw new Error("You aren't allowed!");
+
+    const hasPermissions = ctx.request.user.permissions.some(per =>
+      ["ADMIN", "ITEMDELETE"].includes(per)
+    );
+
+    if (ownsItem || hasPermissions) {
+      return ctx.db.mutation.deleteItem({ where }, info);
+    } else throw new Error("You aren't allowed!");
     // 3. Delete
-    return ctx.db.mutation.deleteItem({ where }, info);
   },
   async signup(parent, args, ctx, info) {
     // 1. lowercase email arg
@@ -159,6 +169,31 @@ const Mutations = {
     });
     // 8. Return user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // 1. Check for logged in
+    if (!ctx.request.userId) throw new Error("You must be logged in!");
+    // 2. Query current user
+    const currentUser = await ctx.db.query.user(
+      { where: { id: ctx.request.userId } },
+      info
+    );
+    // 3. Check if they have permissions
+    hasPermission(currentUser, ["ADMIN", "PERMISSIONUPDATE"]);
+    // 4. Update user
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions
+          }
+        },
+        where: {
+          id: args.userId
+        }
+      },
+      info
+    );
   }
 };
